Reorganized some parts of project
Move parts of Actor into Actor folder. Added State class
