+++ Source/Oshima/Actor/ActorType.cpp
