+++ Source/Oshima/Actor/ActorType.h
