+++ Source/Oshima/State/State.cpp
@@ -1,0 +1,21 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "State.h"
+
+UState::UState()
+{
+}
+
+void UState::Activate()
+{
+	isActive = true;
+}
+
+void UState::DeActivate()
+{
+	isActive = false;
+
+}
+
+
